Fixed Vivecharacter not picking up base simulating actors anymore (when changing for the new IsHeld interface function I set a node wrong)
Added the plugins new check for destructible components being fractured, not entirely sure that I should even be doing this.
In the future I may force the end user to properly clean up destructible meshes instead by dropping them or attaching them to
a root component that is gripped instead.