Moved more of the FSavedMove logic into the base FSavedMove of BaseVRCharacterMovementComponent
Corrected an issue where step up functions were not correctly re-winding failed stepup attempts
causing stuttering and jitter on step up attempts in multiplayer and solo at times.

Now using a more correct movement rewind for when the server sends location corrections to the clients
in VR.

Now rounding climbing movement to 2 decimal precision in multiplayer so that it matches
the resolution that is sent to the server.

Now using a custom quantized FTransform for the grip transforms over network.
This transform is only 2 decimal precision so in the blueprint when checking for
the slot grip location I am now only checking up to 2 decimals of precision for a match

I would like to start doing the interaction settings as a quantized value as well but it was clearing
stored values in the editor and I didn't want to ruin peoples currently working setups.