Local only grips now replicated to non owning clients to take the workload off of the users. Client still has total authority over the initial grip so it is instant.
Renamed LocalOnly_Not_Replicated to ClientSide_Authoritive

Added gameplay tag interface to each grippable component/actor, allows for extremely
flexible object management without enforcing gameplay systems. Used in the example
for defining if an object drops with specific buttons and a few other things.

Now that gameplay tags are an official part of the engine they should be first class
citizens in the plugin.

Removed the ObjectType function/variable from the VRGripInterface, close to useless when
compared to the gameplay tags, I don't know if anyone was using it anyway.

Modified all template objects to use the gameplay tags for drop buttons an secondary grip
drop buttons.

Modified the blueprints in the vr character to support the gameplay tags and cleaned them
up in general.

Fixed a packaging error with the make / break functions for the Transform_NetQuantize struct.

Added DefaultGameplayTags.ini to repository, this contains the generic default tags currently
in use in the template project.