Fixed console drop settings
Fixed thirdperson locomotion head, added new variable to plugin replicated camera
that when set will set the relative position of the camera during tick instead
of during the camera view.

(Camera by default doesn't set relative location to HMD when not the current view
target. This allows it to continue to set the location when focus is on a different
camera).