Adjusted template navmesh to be better (had left it at epic defaults and it needed a refresh pass)
Fixed a missing Super:: call in PhysCustom so that the HandleCustomMovement event would
fire for blueprint users on custom movements that are not already handled by the plugin.

Expanded EVRConjoinedMovementModes enum to include 6 custom movement mode definitions
this will make it easier to select a custom movement mode that doesn't overlap with already used
values. This also allows SetReplicatedMovementMode to work with those 6 custom movement mode
index's for true during movement mode switching without explicit RPCs.

Changed the auto step up for climbing to call SetReplicatedMovementMode now instead of
directly changing movement mode, this allows it to still function if the server rolls
back a move and replays (previously in rare cases it was hanging on climbing as the rollback replay
wasn't re-setting the out of climbing mode).