Switched 3Dkeyboard button component over to the VRButtonComponent c++ class as its base class
Fixed a bug with scaled meshes and the VRButtonComponent-Toggle_Stay mode.

Fixed a naming occlusion with the ValidOverlap detection in the VRButtonComponent

Added a bIsEnabled property to the VRButtonComponent so that users can disable it.

Added SteamRequestGroupOfficers function to the sessions plugin.

Added GetSteamGroups to the sessions plugin.

Added GetSteamFriendGamePlayed function to the sessions plugin.