Removed SteamVRHMD includes and SteamVR/Private from being referenced entirely, it is no longer needed.
Had to copy some matrix conversion functions to the OpenVRFunctionLibrary but the
end result is for the better.

Added dependancy modules to the plugin modules as 4.17 now allows for it

Removed all kismet library references, they were temp to begin with to allow for testing a long
time ago.

Lever component was missing some header includes.