Removed MotionControllerLocDelta from TickGrip
Not only can you manually calc it (and probably should) with custom grips.

But being in worldspace the offset could be wrong due to character movement.

Better to let items calc it themselves and store the variables for more advanced use instead.
I don't think it had much use in its current state.