Made "Light" versions of the RPCs that send saved movements to the server. These do not pass in the RequestedVelocity and CustomVRInputVector. They are called if both the RequestVelocity and CustomVRInputVectors are zero. This cuts down on some of the additional overhead built into the replication in the VR characters.
Also the saved movement capsule Yaw value being sent over the network is now the same calculated Pure yaw that the GenerateOffsetToWorld
uses. This will make running back movements slightly faster, as well as slightly more accurate as passing
in just the normal HMD yaw before could yeild slightly different results when the camera is rotated off axis than
if calculating entirely off of the full rotator.

I also went into the DropActor RPC in the blueprints for VivePawnCharacter and quantized the requested
angular and linear velocities as well as only passing in the object now instead of the grip (which grip
is not reconstituted on the other side by getting grip by object).
This cuts the bytes used to replicate this RPC to 1/5 of what it was.


*Car has all 4 seats filled in now.*