Commented out the VRSceneCaptureComponent
It is incomplete and I am unsure if I will ever finish it, however I wanted to leave
the source behind for people that may want to make one themselves.

Never got around to adding additional captures to it, but it already handles some
stereo specific things and uses the latest possible game thread positional update.