Added Euro Low Pass Filter to the plugin (same on that Epic is using with the VR Editor for their laser beam smoothing). It has ResetEuroFilter and RunEuroFilterSmoothing nodes in the VR Expansion Library to control it.
Also added an auto arc calculation node to the VR Expansion Function Library so that the
arc calculation and setting is faster and clean and is the same as the VR Editor.

Modified the laser beam to be settable between normal linear beams and the VR Editor style
smoothed "wobble" beams.