Changing bIgnoreHandRotation to reconstruct and resolve the controller instead of requiring it be modified prior to passing to the server.
Changed GetControllerRelativeTransform to no longer take an input object for
checking bIgnoreHandRotation, it is no longer needed. Kept the node around anyway
because the default in engine nodes variables are named silly and confusing.

Changed the door to use a force constraint instead of an acceleration one.
This makes it not try to fight on its hinges.