Added cvar to turn on COM visualization for grips "vr.DrawDebugCenterOfMassForGrips"
Removed some logic checks from the BP Vive character so that it wouldn't filter out
destructible objects by default on grip. Now it just checks for simulation and also
always passes in the bone name for objects if they don't implement gameplay tags.


A better destructible would still be to create one with the interface and auto drop it on
fracture.