Removed the Interactible.Door/Lever/Dial gameplay tags, they aren't needed for the template.
Rather people define their own categories if they use them for that.


Removed UseSeperateTouchSockets from the gameplay tags and from being used in the BP character.
Now that BP profiles exist I would rather that they be used instead as they do this in a better
manner anyway.

Set the FPS test pawn to the Oculus_Touch profile so that objects align correctly.