Converted physics grips to use the same pathway regardless of type (no longer a seperate logic branch for manipulation grips, now they just use different defualts).
Removed the bDoNotSetCOM option from the advanced physics settings.

Replaced it with an enum with a bunch of seperate options that also now apply to the
manipulation grip.

This has the side effect of slightly lowering memory (1 transform) and slightly lowering
CPU costs of physics grips (less transform logic and checks) as well as greatly improving the
flexibility of the physics grips.