Made a barebones VRCharacter that removes the motion controllers and parent relative attachment.
Can add your own to the character then, however for motion controllers you should call
InitForVRCharacter on them to set the correct tick ordering and any other setup I decide
to add in the future.

Keep in mind that doing this overrides the minor network optimization I have for the controllers
where they can use the characters replication of their position instead of their own (it is very minor).