Added grippable actors to the grippables prefabs.
Lets you change out the root component at will to use skeletal or static mesh components.
This is useful for making a base class and changing out the type of root component further down.
It was always possible to do in BP, but this makes it easier by pre-defining it.

I suggest using the plugins OptionalRep(Static/Skeletal)MeshComponents as the roots
as they implement the optional movement replication and you will be missing the
StaticMeshComponentReplicatesMovement option that StaticMeshActor adds in.