Added a few new functions to controller profiles
Made the transforms in the controller profiles FTransformNetQuantize so that they
are smaller to send as well as able to be net compared (defualt transforms have too much
precision and can trigger OnRep multiple times because it doesn't detect them as
being the same value).

(Profiles do not replicate the button mappings, only the transforms)

Cleaned up some of the node for the controller profiles and made the client send to the server
what profile it loaded, so that the server can send to remote clients as well.