!Important! Bug fix
It appears that the engines native TArray replicatio is in capable of correctly
initializing non replicated variables of an index when an index is removed and
replaced on the same serializing frame.

This causes the repped values to correctly transfer over, but the non repped ones
to remain stale and un inited.

I am now comparing the last known GripID to the current one when a grip is
replicated over, if they do not match I am wiping out the non replicated variables
to their empty values and letting the grip init cleanly.

I will NEED to report this to Epic, it appears to be a deeply nested and unknown
issue with how the TArray replication is handled.

It is fairly terrifying that this can happen considering how much it can break
things and that it could effect every multiplayer project using the engine.