Adding in bIsForTeleport to the grip script GetWorldTransform, useful thing to know.
Also Added bForceDrop to the grip scripts, when true the grip will be automatically
dropped on the first valid and safe point. Can be set with ForceGripToDrop() function.

Anything more refined will likely required bindable events or something to be made in
your custom GripScript for the owner to handle on demand.