Fixed some seated mode issues
Am now forcing attachment on the simulated client side as well just in case.

Also I am clearing some movement prediction data so that the server doesn't
wait for a few seconds for a valid timestamp move in an old out of date range.

Also I am network rounding the stored location for seated mode and the stored
yaw for seated mode. This prevents an initial movement correction on leaving the
seat in some situations.