Made virtual stock settings part of the global settings structure
this makes it easier to store a copy in the character and replicate it to the server
as well as from the server to the gripped object for setting.

Also added the option to UseGlobalVirtualStockSettings in the gun tools grip script.

When enabled it will load the current settings from the global settings (default true and only for locally controlled).

When disabled it uses local settings.