Added some new Virtual Stock features to GunTools
Added an event for when the virtual stock is enabled / disabled with the distance based
setting.
OnVirtualStockModeChanged

Added StockSnapLerpThreshold, when within the StockSnapDistance will lerp across this
threshold from 0-1.0 for influence so that the transition is smoothed out.
If it is set to 0.0f then the transition will be a hard snap instead.

Added StockLerpValue, you can read this from the VirtualStockSettings var
in the grip script to get the current state of that lerp if you need it.

Also added an event to the LerpToHand GripScript
OnLerpToHandFinished