Revised the LerpToHand grip script to work a bit differently.
LerpSpeed is not LerpDuration and is the number of seconds that you want the lerp to take.

LerpMinSpeed is a new var that defines the minimum speed in UU per second that you want
the object to have (from intial distance on grip to the hand across the LerpDuration).
It scales the LerpDuration in order to maintain this minimum speed.

LerpMaxSpeed is the same as above except for maximum speed.

The lerp curve now when used, only uses the 0.0 - 1.0 range of the curve and it maps
the above variables to this curve.