Adjusted pre physics grip scripts so that they correctly set the Com again after changing properties.
Made the guntools pivot offset also apply to the center of mass on grips that set it,
and the relative position.

Added the same orientation component setup that the melee script uses to re-align
the constraint and the mesh, the pivot offset applying to the center of mass REQUIRES
an orientation component to be set and used so that I have a valid orientation to
work off of.