added viewpoint setting for grippable AI characters
added viewpoint setting for grippable AI characters, can alter the value of the "ViewOriginationSocket" property and it will use that socket world position when querying GetActorsEyesViewPoint instead of the default of the capsule center / facing.

This will let you move the vision detection for AI Sight Sense and the like to the actual head rotation / facing.