Grip Fixes
Grip Fixes

Fixed client auth grips initiated on the server side not setting their rep cache variables correctly (throwing incorrect events for the server).

Also fixed secondary grips that already exist when a grip loads in not throwing the OnSecondaryGrip event to new players, this was introduced when the grip caching was overhauled in 4.25.