Merge pull request #49 from ronaldburns/patch-1  - Change VRBaseCharacter components to Optional subobjects
Merge pull request #49 from ronaldburns/patch-1
Change VRBaseCharacter components to Optional subobjects

This is totally not safe with the entire code base assuming that the components exist. There WILL be code at some point that crashes out because of this, even if currently I have everything fully locked behind checks.

That being said, it has to be specifically done by something looking to do it, and can't be an accidental issue. So I don't really see the harm in letting people shoot themselves in the foot and seeing how far they get ;p.