this may resolve is object tracked issues
this may resolve is object tracked issues, needs testing. Its a problem because Right Grip and Left Grip are supposed to map to Left/Right everywhere correctly. But we already know that tracking is weird if selecting one or the other and the pose appears different. Epic may have something wrong here.